build: do not redownload if packages already exist

+++ build.sh
@@ -50,28 +50,36 @@ cd "$DL"
 
 echo " - Python"
 
-wget -Nc "$PYTHON_URI"
+if [ -e "$PYTHON_PKG" ]; then
+    echo "Using existing $PYTHON_PKG"
+else
+    wget -Nc "$PYTHON_URI"
+fi
 
 echo " - libffi"
 
-# get a JSON with an anonymous token
-token=$(curl -s "https://ghcr.io/token?service=ghcr.io&scope=repository%3Ahomebrew/core/go%3Apull" | jq -jr ".token")
+if [ -e "$LIBFFI_PKG" ]; then
+    echo "Using existing $LIBFFI_PKG"
+else
+    # get a JSON with an anonymous token
+    token=$(curl -s "https://ghcr.io/token?service=ghcr.io&scope=repository%3Ahomebrew/core/go%3Apull" | jq -jr ".token")
 
-digest=$( \
-    curl -s \
-    -H "Authorization: Bearer ${token}" \
-    -H 'Accept: application/vnd.oci.image.index.v1+json' \
-    $LIBFFI_MANIFEST_URI \
-    | jq -r '[.manifests[] |
-            select(.platform.architecture == "arm64"
-            and .platform."os.version" == "'"$LIBFFI_TARGET_OS"'")
-        ] | first | .annotations."sh.brew.bottle.digest"' \
-)
+    digest=$( \
+        curl -s \
+        -H "Authorization: Bearer ${token}" \
+        -H 'Accept: application/vnd.oci.image.index.v1+json' \
+        $LIBFFI_MANIFEST_URI \
+        | jq -r '[.manifests[] |
+                select(.platform.architecture == "arm64"
+                and .platform."os.version" == "'"$LIBFFI_TARGET_OS"'")
+            ] | first | .annotations."sh.brew.bottle.digest"' \
+    )
 
-curl -L -o "$LIBFFI_PKG" \
-    -H "Authorization: Bearer ${token}" \
-    -H 'Accept: application/vnd.oci.image.index.v1+json' \
-    "$LIBFFI_BASE_URI/sha256:$digest"
+    curl -L -o "$LIBFFI_PKG" \
+        -H "Authorization: Bearer ${token}" \
+        -H 'Accept: application/vnd.oci.image.index.v1+json' \
+        "$LIBFFI_BASE_URI/sha256:$digest"
+fi
 
 echo "Building m1n1..."
